backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Pydantic / settings
pydantic[email]==2.10.3
pydantic-settings==2.7.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Redis / Celery
redis[asyncio]==5.2.1
celery[redis]==5.4.0
kombu==5.4.2
celery-redbeat==2.2.0
flower==2.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# HTTP client
httpx==0.28.1
tenacity==9.0.0

# Excel import
openpyxl==3.1.5
chardet==5.2.0

# Observability
loguru==0.7.3
prometheus-fastapi-instrumentator==7.0.0
prometheus-client==0.21.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
